examples/stt_assemblyai_transcription/pyproject.toml
[project]
name = "getstream-stt-assemblyai-transcription-example"
version = "0.1.0"
description = "Example project showing how to transcribe a call using STT with AssemblyAI"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "getstream[webrtc]>=2.3.0a0",
    "getstream-plugins-assemblyai>=0.1.0",
    "python-dotenv>=1.1.1",
]
